Split delta logic and ddb load
